Implemented getBalance and getMarket functions
Implemented getMarket function to return the values of all the currencies available on the market.
Implemented getBalance function to return the quantity and the value of the currency in client's portfolio.
Added a Constructor of Crypto to be able to utilize getPrice; Crypto::getPrice initialization at a random value.
Modified HandleBuy/Sell consequently; new conditions on the client's cash portfolio balance, update its balance for each transaction.
Added all the initialized client balances in Server.h.
Modified Client behavior to test all the functionnalities.